Add Variable and Coeff
