+++ runtime/src/conf_transfer.rs
@@ -319,11 +319,11 @@ mod tests {
     #[test]
     fn test_call_function() {
         with_externalities(&mut new_test_ext(), || {
-            let proof: [u8; 192] = hex!("b4ca09c7c2d9e887e40c37e70a56d97a41ea8da03e40b253e1933f1677c7309c4d3ab8d5aeb9907bf7b328df12af66c3a01c52cf40e9cbcf4f227cab510458835064867a4bbad789be83df0fa62aace268d30296ab49606a75d5bb2f2a4a55431323565a7a3fe94f21c34c71d8aa1a05a7efb5e4c04766ed9920a8c16e8f6fd92c5fec20673bd7561ed59f122239ffd1ac2638fd2d223d0520179a0c2b0fb875ebcef2925782c9f5496ddcf121e012f97f3be9bd3cc74a037064685554e367a7");
+            let proof: [u8; 192] = hex!("91e1e662f17b871026431a9c9b48d4bf3a03aa4f8740a841252583d4472f42195d8b8cc1af8c867f63efb84d2d7b8392b26cfed48591f6128a8e1d43b9f4479c5ddb034e01035f192f7f9627d0525280a646f780c71e66286fed3f838a134e2613d311f6990ec248051ac7e648b8afcabc8924242392afe4ed97eee5b6df6692e54d9d65e8e19eda2b4f3148694baced8fa508a2c9d72c3fe2849158b3dac922ae716f259aab059c503a31a1f81e3836ea6e90de8dfa828f6a0384595995a78c");
             let pkd_addr_alice: [u8; 32] = hex!("fd0c0c0183770c99559bf64df4fe23f77ced9b8b4d02826a282bcd125117dcc2");
             let pkd_addr_bob: [u8; 32] = hex!("45e66da531088b55dcb3b273ca825454d79d2d1d5c4fa2ba4a12c1fa1ccd6389");
-            let enc10_by_alice: [u8; 64] = hex!("7db0710337156f353f5c03e41a1b7bfda913e987770c58908da7080d28b0403fdff1316f7cd3189e73563a5481bc953eb3796fb5565f433f3acdbe128859326b");
-            let enc10_by_bob: [u8; 64] = hex!("8d517bf8f8f2aab7090eaa7a57d6ddf9930d8b7198e1b980f1c977255234fca6dff1316f7cd3189e73563a5481bc953eb3796fb5565f433f3acdbe128859326b");
+            let enc10_by_alice: [u8; 64] = hex!("661550a1a5ac8d297a0bfd5eb2f8f8d9fb03a2c694e265d36e11aa015bb5c49cf07b68ee25b21ea9e6e8346e963e9e55365a1f5880dfc38caba4dd54c26f25c7");
+            let enc10_by_bob: [u8; 64] = hex!("aae4a58760dcab7646df25e4cf45e02af373f2eabb8fd47b94ed7bf2f7e488b5f07b68ee25b21ea9e6e8346e963e9e55365a1f5880dfc38caba4dd54c26f25c7");
             let rvk: [u8; 32] = hex!("f539db3c0075f6394ff8698c95ca47921669c77bb2b23b366f42a39b05a88c96");
 
             assert_ok!(ConfTransfer::confidential_transfer(
